Move the test for build_embedded_model to test_embed_utils.py where its friends live.
